API/requirements.txt
Flask==1.1.2
numpy==1.17.2
nltk==3.5
tensorflow-cpu==2.2.0
gunicorn==20.0.4